nevermind about /var/lib/condor-ce/history
Per Derek, we want PER_JOB_HISTORY_DIR to match DataFolder,
and for the hosted-ce we can leave CondorCEHistoryFolder unset.

Well, it turns out that you can't really unset it, because since
SOFTWARE-2463, there is always a default for it - both in the
packaged ProbeConfig and the gratia common code.  So anyway for
now we'll just leave it at that default.